arquivos do front refatorados e integrados com minio e bd para a funcao upload
